Introduce factories for the Expenses seeds
- added a factory for the Expense model
- minor refactoring to the seeds scripts
